# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Gaming Session Planner",
    "description": "Use Github login to plan video gaming sessions",
    "version": "1.0.0"
  },
  "host": "cse-341-project-2-2548.onrender.com",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "JWT": {
      "type": "apiKey",
      "in": "cookie",
      "name": "jwt",
      "description": "JWT token stored in cookie"
    }
  },
  "paths": {
...
    "/users/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
  "security": [
    {
      "JWT": []
    }
  ]
}